+++ TransferScreen.h
@@ -53,7 +53,7 @@ private:
     sf::Font& fontRef;
 
     std::vector<PlayerRow> playerRows;
-    int hoverIndex = -1;
+    size_t hoverIndex = -1;
     const float LIST_START_X = 50.0f;
     const float LIST_START_Y = 100.0f;
     const float ROW_HEIGHT = 35.0f;
